+++ ntex-io/src/lib.rs
@@ -202,8 +202,8 @@ pub enum IoTaskStatus {
 /// I/O status update events.
 #[derive(Debug)]
 pub enum IoStatusUpdate {
-    /// Keep-alive timeout has occurred.
-    KeepAlive,
+    /// The dispatcher timer has expired.
+    Timeout,
     /// Write backpressure is currently active.
     WriteBackpressure,
     /// The connection is no longer usable.
@@ -267,7 +267,7 @@ mod tests {
 
     #[test]
     fn test_fmt() {
-        assert!(format!("{:?}", IoStatusUpdate::KeepAlive).contains("KeepAlive"));
+        assert!(format!("{:?}", IoStatusUpdate::Timeout).contains("Timeout"));
         assert!(format!("{:?}", RecvError::<BytesCodec>::KeepAlive).contains("KeepAlive"));
         assert!(
             format!("{:?}", RecvError::<BytesCodec>::WriteBackpressure)