+++ ntex/src/http/h1/dispatcher.rs
@@ -476,14 +476,14 @@ where
                 return Poll::Pending;
             };
             match status {
-                IoStatusUpdate::KeepAlive if self.timers.active.is_write() => {
+                IoStatusUpdate::Timeout if self.timers.active.is_write() => {
                     if let Err(err) = self.write_timer_expired() {
                         Poll::Ready(self.ctl_peer_gone(Some(err)))
                     } else {
                         Poll::Pending
                     }
                 }
-                IoStatusUpdate::KeepAlive => Poll::Pending,
+                IoStatusUpdate::Timeout => Poll::Pending,
                 IoStatusUpdate::WriteBackpressure => {
                     self.timers
                         .start_write(&self.io, self.codec.cfg.write_timeout);
@@ -626,7 +626,7 @@ where
             }
             // the timer is reported through the status update
             match self.io.poll_status_update(cx) {
-                Poll::Ready(IoStatusUpdate::KeepAlive) => {
+                Poll::Ready(IoStatusUpdate::Timeout) => {
                     if let Err(err) = self.write_timer_expired() {
                         return Poll::Ready(Err(err));
                     }